workspace: Detect git worktrees as "colocated workspaces"
Previously, only the default workspace could be
colocated. This adds logic to the colocation-detection
code to support finding a worktree as well as a full-on
.git directory.

Secondly, in a couple of places, this discovered worktree
is opened instead of always opening the backing repo. This
is so that when you import HEAD, you import it from the current
worktree. This work can't be finished until we can store multiple
git HEADs in the view, but is continued in the following few diffs.