view: deprecate git_head, introduce git_heads (a HEAD per workspace)
Previously, JJ's view of the git head (HEAD@git) could only track one
HEAD, for only one workspace, specifically the one colocated with the
main .jj/repo.

Since we now allow workspaces to be colocated, we need to track a
separate git HEAD per workspace. This allows us to import HEAD
for only the current workspace, and export HEAD to the current workspace
when @- moves. Thereby allowing git tools to behave correctly in those
workspaces.

This affects a lot of stuff. And a lot of tests. A lot more things need
to know the current workspace ID, including e.g. the revset resolvers,
 and the commit templater.