git: implement git_worktree_remove
We need this for when we forget colocated workspaces, so that users
do not have to `git worktree remove` manually.

The validation / removal separation is so we can two-phase-commit the
removals during a multi-workspace forget operation, checking they're all
valid and able to be removed during the transaction and removing the
worktrees afterwards. Missing worktrees (like when the user has ALREADY
run `git worktree remove`) are ignored.