docs: mention that that Rust 1.76 is required to install from source
We have our MSRV recorded in `Cargo.toml`, but at least with Rust
versions < 1.71, the build fails before the MSRV is checked (IIUC),
and you get a cryptic message like this:

```
error: Package jj-lib v0.22.0 (<some path>) does not have feature git2.
```

The message is clearer from 1.71, but that doesn't help users who have
an older version. Let's mention the MSRV everywhere we tell the user
to run `cargo` in the installation instructions.