cli: make most command aliases not interfere with completion
This follows up on fd271d3, making most command aliases hidden as far as
`clap` is concerned and adding them to the docstring instead. This means
that they will not interfere with shell command-line completion.

In the process, I normalized the format of the first docstring line a
bit.

A downside is that the alias definition is now often in a different file
than the documentation for this alias definition.  I considered either
moving the alias definition to the same file as the docstring where the
docs for the alias go, or the other way around, but it didn't seem to
work easily.  Moving `#[command(alias("b"))]` next to `BookmarkCommand`
would make the alias conflict with the deprecated `jj branch` command
that also uses BookmarkCommand. Moving the whole docstring for `bookmark
list` into `BookmarkCommand` would put it far for the (potential future)
docstrings for the options to `bookmark list`. So, I gave up on this.