workspace: Reload the repo from disk in `workspace add`
This allows us to have a differently configured GitBackend in the newly-
added workspace, i.e. one which was opened at the workspace root if we
are creating a colocated workspace.

This was easier than I anticipated -- we can use RepoLoader by
itself to load the store from disk, and then create the new working
copy commit immediately.