cli: add `jj git colocate`
Initially, we can only colocate the primary workspace.

See <https://martinvonz.github.io/jj/latest/git-compatibility/#converting-a-repo-into-a-co-located-repo> for the steps.

Adds a Workspace::reload() function to reload the workspace from the
file system, which is technically not necessary for this diff, but will
become necessary soon, when we make the GitBackend colocation-aware.