docs: add top-level CONTRIBUTING.md
It just links to our actual docs and repeats the passage from README.

I find that I often automatically scan for a `CONTRIBUTING.md` file in
the repo and get annoyed when it's hard to find. This should be helped
by the previus commit, but this will hopefully go further in putting
everyone on a road leading to Rome as quickly as possible.

This will also hopefully mean that more people read the instructions on
our website rather than GitHub, which will be a somewhat nicer
experience after #4822.